requirements.txt
flask
requests
python-dotenv
mysql-connector-python
